refactor: move newsletter handling to a separate class
